success seperate enum
