+++ src/main/java/softeer2nd/chess/pieces/piecetype/Bishop.java
@@ -1,7 +1,10 @@
 package softeer2nd.chess.pieces.piecetype;
 
+import softeer2nd.chess.pieces.piecetype.enumutils.Type;
+
 import static softeer2nd.chess.pieces.piecetype.Piece.Color.*;
-import static softeer2nd.chess.pieces.piecetype.Piece.Type.BISHOP;
+import static softeer2nd.chess.pieces.piecetype.enumutils.Type.*;
+
 
 public class Bishop extends Piece {
     private Bishop(Color color) {
